refactor: Move more utilities to defence module
